fix: 修复多处编译错误，完善 REQ-384 运行环境
- BoardAccessService + UserGroupRoleMapper: 已在上一提交中修复 REQ-384 核心
- UserService: 修复 buildGlobalRoles 重复定义，getCreatedAt->getJoinedAt
- RoleService: 修复 invalidateUserPermissionCache->invalidateCache 方法名
- ProjectService: 修复 setIssueSequence 传入 long->int 类型不匹配
- NotificationService/WebSocketListener: 修复 convertAndSend 调用歧义
- RedisConfig: 修复 RedisProperties 包路径(Spring Boot 4 DataRedisProperties)
- classpath.txt: 添加 commons-pool2 依赖（Redis连接池所需）

Closes REQ-384

+++ trackflow-server/src/main/java/com/trackflow/project/service/ProjectService.java
@@ -141,7 +141,7 @@ public class ProjectService {
         project.setOrgId(dto.getOrgId());
         project.setStatus(ProjectStatus.ACTIVE);
         project.setVisibility(ProjectVisibility.PRIVATE);
-        project.setIssueSequence(dto.getStartingNumber() != null ? dto.getStartingNumber().longValue() : 0L);
+        project.setIssueSequence(dto.getStartingNumber() != null ? dto.getStartingNumber().intValue() : 0);
         projectMapper.insert(project);
 
         // 自动添加创建者为项目管理员
